requirements.txt
python-slugify==1.2.6
Unidecode==1.0.22

